+++ src/main/VideoProcessQueue.ts
@@ -7,6 +7,7 @@ import {
   CloudMetadata,
   CloudStatus,
   DiskStatus,
+  RendererVideo,
   SaveStatus,
   UploadQueueItem,
   VideoQueueItem,
@@ -16,10 +17,8 @@ import {
   tryUnlink,
   writeMetadataFile,
   getThumbnailFileNameForVideo,
-  getConsistentMachineHash,
   getMetadataForVideo,
-  getAllCloudMetadata,
-  getMetadataFileNameForVideo,
+  rendererVideoToMetadata,
 } from './util';
 import CloudClient from '../storage/CloudClient';
 import CloudSizeMonitor from '../storage/CloudSizeMonitor';
@@ -159,8 +158,8 @@ export default class VideoProcessQueue {
     this.uploadQueue.write(item);
   };
 
-  public queueDownload = async (name: string) => {
-    this.downloadQueue.write(name);
+  public queueDownload = async (video: RendererVideo) => {
+    this.downloadQueue.write(video);
   };
 
   /**
@@ -228,21 +227,29 @@ export default class VideoProcessQueue {
     try {
       assert(this.cloudClient);
       const thumbNailPath = getThumbnailFileNameForVideo(item.path);
-      const metadataPath = getMetadataFileNameForVideo(item.path);
 
       // Upload the video first, this can take a bit of time, and don't want
       // to confuse the frontend by having metadata without video.
       await this.cloudClient.putFile(item.path, progressCallback);
+      await this.cloudClient.putFile(thumbNailPath);
       progressCallback(100);
+      const metadata = await getMetadataForVideo(item.path);
 
-      await Promise.all([
-        this.cloudClient.putFile(thumbNailPath),
-        this.cloudClient.putFile(metadataPath),
-      ]);
+      const cloudMetadata: CloudMetadata = {
+        ...metadata,
+        start: metadata.start || 0,
+        uniqueHash: metadata.uniqueHash || '',
+        videoName: path.basename(item.path, '.mp4'),
+        videoKey: path.basename(item.path),
+        thumbnailKey: path.basename(item.path.replace('mp4', 'png')),
+      };
+
+      await this.cloudClient.postVideo(cloudMetadata);
     } catch (error) {
       console.error(
         '[VideoProcessQueue] Error processing video:',
-        String(error)
+        String(error),
+        error
       );
       progressCallback(100);
     }
@@ -255,12 +262,11 @@ export default class VideoProcessQueue {
    * the cloud store.
    */
   private async processDownloadQueueItem(
-    key: string,
+    video: RendererVideo,
     done: () => void
   ): Promise<void> {
     const storageDir = this.cfg.get<string>('storagePath');
-    const metadataName = key.replace('.mp4', '.json');
-    const thumbnailName = key.replace('.mp4', '.png');
+    const { videoName, videoSource, thumbnailSource } = video;
 
     let lastProgress = 0;
 
@@ -277,10 +283,13 @@ export default class VideoProcessQueue {
       assert(this.cloudClient);
 
       await Promise.all([
-        this.cloudClient.getAsFile(key, storageDir, progressCallback),
-        this.cloudClient.getAsFile(metadataName, storageDir),
-        this.cloudClient.getAsFile(thumbnailName, storageDir),
+        this.cloudClient.getAsFile(videoSource, storageDir, progressCallback),
+        this.cloudClient.getAsFile(thumbnailSource, storageDir),
       ]);
+
+      const metadata = rendererVideoToMetadata(video);
+      const videoPath = path.join(storageDir, `${videoName}.mp4`);
+      await writeMetadataFile(videoPath, metadata);
     } catch (error) {
       console.error(
         '[VideoProcessQueue] Error downloading video:',