+++ src/main/types.ts
@@ -199,7 +199,7 @@ type Metadata = {
   mapID?: number;
   challengeModeTimeline?: RawChallengeModeTimelineSegment[];
   soloShuffleTimeline?: SoloShuffleTimelineSegment[];
-  level?: number;
+  keystoneLevel?: number;
   encounterName?: string;
   protected?: boolean;
   soloShuffleRoundsWon?: number;
@@ -210,6 +210,18 @@ type Metadata = {
   tag?: string;
   delete?: boolean; // signals video should be deleted when possible
   uniqueHash?: string; // used for cloud video grouping
+};
+
+/**
+ * We mandata some fields are present for cloud videos that are optional for
+ * disk based videos.
+ */
+type CloudMetadata = Metadata & {
+  videoName: string;
+  videoKey: string;
+  thumbnailKey: string;
+  start: number;
+  uniqueHash: string;
 };
 
 /**
@@ -230,7 +242,7 @@ type RawCombatant = {
  * add when reading the file.
  */
 type RendererVideo = Metadata & {
-  name: string;
+  videoName: string;
   mtime: number;
   videoSource: string;
   thumbnailSource: string;
@@ -416,6 +428,7 @@ type CloudObject = {
 interface ICloudClient {
   list: () => Promise<CloudObject[]>;
   delete: (key: string) => Promise<void>;
+  deleteVideo: (videoName: string) => Promise<void>;
 }
 
 interface IBrowserWindow {
@@ -478,4 +491,5 @@ export {
   ICloudClient,
   IBrowserWindow,
   UploadQueueItem,
+  CloudMetadata,
 };