+++ src/renderer/Layout.tsx
@@ -66,7 +66,7 @@ const Layout = (props: IProps) => {
         videoFilterQuery: '',
         page: Pages.None,
         category: newCategory,
-        selectedVideoName: first?.name,
+        selectedVideoName: first?.videoName,
         playingVideo: first,
         numVideosDisplayed: 10,
       };
@@ -93,7 +93,7 @@ const Layout = (props: IProps) => {
         icon={
           <Badge
             badgeContent={numVideos}
-            max={999}
+            max={9999}
             style={{ transform: 'translate(15px, 5px)' }}
             sx={{
               '& .MuiBadge-badge': {